Update patch set 4
Patch Set 4:

(11 comments)

Harman, could you please add a test where the NormalizedNodeStreamWriter is used with a Serializable. So essentially this is what you need to do. Create a new class which implements Serializable. This class contains a NormalizedNode as a field. You should be able to successfully serialize and deserialize an instance of the class using the stream reader/writer.

Patch-set: 4
Label: Verified=0