Update patch set 5
Patch Set 5:

(1 comment)

Actually there's a 4th choice above. Similar to #2 - ditch the enums but leave the original enums alone and define new classes with byte fields.

The test yang model used by the unit test defines some of the possible yang constructs but we need to extend it to test all the possible yang constructs. We should strive for 100% line and branch coverage in the stream reader and write classes.

Patch-set: 5