Update patch set 2
Patch Set 2:

Ok - I see the point with breaking consistency - didn't think about that. With that issue, I don't think it's possible to do this (unless the broker's new*Transaction APIs are changed to take the logical data store). I'll drop these changes.

I can make changes in the clustered data store to minimize the overhead of unused data store transactions.

However, I should point out that the clustered data store does currently break consistency in this manner. A TransactionProxy is initially created in each data store, as instructed by the broker, but the actual underlying shard transactions are lazily created as read/modification operations are performed by the client. It's done this way because we can't create a shard transaction until an operation that specifies a path is performed, otherwise we don't now which shard in which to create a transaction, unless we automatically initially create a "blank" transaction in every shard which wouldn't be ideal.  This would be alleviated if the broker's new*Transaction APIs were changed to take the logical data store.

Patch-set: 2