Update patch set 2
Patch Set 2:

Ok, I see, but the point you raised about Clustered Data Store needs to be release noted (ideally in wiki) + open bug tracking this issue so we can point people to known bugs and maybe reference it as a source for bugs which may be exposed by this behaviour for clustered data store.

I am asking for this bug, to make sure we will NOT forget about clustering behaviour.

Patch-set: 2