Update patch set 1
Patch Set 1:

So, there is a basic problem with that - there are *NO* current junit tests around neutron NB at all - so asking for them for this case is a barrier to merge that has not existed previously.  Rather, there should be a defect to add junit test to the whole service.

Patch-set: 1
Reviewer: Gerrit User 590 <590@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0