BUG-2255: introduce PingPongDataBroker
This is a forwarding DOMDataBroker implementation, which ensures that
there is at most one transaction pending for a particular transaction
chain. This is not useful in all scenarios, but if the users can
restrain themselves to use the supported semantics, they can use it.

Change-Id: I4a9ca12254763829f99e270d6716cd034b4b5ea9
Signed-off-by: Robert Varga <rovarga@cisco.com>
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>
(cherry picked from commit 5764146d24df9d9450ebfbb0aec10cbbcdcc655c)