Update patch set 2
Patch Set 2:

In the ConcensusStrategy , i see election actions and timeouts.

Are we stating that an election is mandatory for which ever concensus strategy we choose?

It might happen that in a 2 node cluster, one has a static configuration that 1 node serves as the primary and other has a failover node. With a strong/eventual consistency. In that case election may not be necessary right.

Patch-set: 2
Reviewer: Gerrit User 1676 <1676@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0