Fix neutron pagination
This patch fixes a logic mistake which results in dead code being
reported. startPos is guaranteed to be non-null by virtue of being
always assigned.

The problem is that the detection that the marker is not found is not
correct, as binarySearch() will report the negated insertion point.

Also improves performance by using a primitive type instead of an
encapsulated object.

Change-Id: I525c2070d2794b1e001465ea925d95d432feca29
Signed-off-by: Robert Varga <rovarga@cisco.com>