Update patch set 4
Patch Set 4:

While adding a default serialVersionUID is a good thing, unfortunately we can't add it for messages that we de-serialize from persistence and/or serialize over the wire as this will break backwards compatibility. For new messages or ones that are safe, I have been adding a default serialVersionUID but we can't do it for many of them at this point.

Not sure what to do here. A lot of classes were touched - would it be possible to revert this patch?

Patch-set: 4
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0