BUG 2335 : Add jolokia feature
This is to make it easy to install jolokia. Earlier it was
possible to install the jolokia bundle the feature just makes
it more convenient.

With this patch I have created a placeholder for extra features. This is
where tiny little utility features of the controller can be put in.

Change-Id: Ifd980656d368902b13834e1e77783151536ca87c
Signed-off-by: Moiz Raja <moraja@cisco.com>