Update patch set 1
Patch Set 1:

Hi Shigeru,

If i look at the flow grouping in opendaylight-flow-types.yang, i do see strict as a boolean leaf element

       leaf strict {
            type boolean;
            default "false";
        }

Default value for this strict flag is false. I believe in this patch, you are kind of hard coding the behavior and FRM will start ignore this flag even if user specify it while making rest/java call. My suggestion is that we should pass this value directly down to openflowplugin without overriding the behavior. Because there might be scenarios where user want to do mass delete using single flow, and this patch won't let them do it.
I my opinion, now the real question is what should be the default behavior of MODIFY/DELETE? I also agree that we should change this default behavior to STRICT as that's the more general scenario people will be expecting. Can you please try changing the default behavior from false to true in the yang module and see if it fixes your issue. If it does, then we should change its default behavior to STRICT, rather then hard coding it in the FRM.

Thanks
Anil

Patch-set: 1