Bug-2277 : DO NOT MERGE - Isolated Leader
This is first set of changes to implement the minority/isolated leader.

A new RaftState has been added and on a scheduler, we check if the leader is isolated, looking at the majority-1 peer statuses.
If Yes, then it switches the behavior to IsolatedLeader.

On the receipt of each AppendEntriesReply, the IsolatedLeader checks for isolation.
And if no, then switches back to either Leader or Follower.

Unit tests are coming.

These changes have been tested with TestDriver and the switching of Leader to IsolatedLeader and back,
on the stopping and reinstating of followers, was successful.

Change-Id: I42b9165cc477d812c7e0e02339537c0f1fe74934
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>