Update patch set 1
Patch Set 1:

Hi Shigeru, Make sense. The only concern is that in FRM we are overriding the behavior that is specified in the yang model, so it's will be good if we can put brief comment in the code to specify why we are doing it. Rest looks good, let me know if once you are done, i will merge this patch.

Patch-set: 1
Label: Verified=0