Bug-2397:Provide a mechanism for stakeholders to get notifications on Raft state change
!!DO NOT MERGE!!

First set of changes of the notification design for review.

In breif, a notifier actor is spawned from the RaftActor. RaftActor notifies the notifier with the role changes.

A sample listener is provided to show how a listener can register with the notifier and get notifications form the notifier.

Tests to follow.

Has been tested with the TestDriver.

Change-Id: I23f16d4e76bb7dae640c544df282293274d9a1cb
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>