Remove lock out of the congested/reuse case
Reusing a transaction is a simple state transition inside the chain,
which helps the system throughput, since more changes will be packed in
a single transaction.

This patch optimizes this reuse as a fast path which is performed
without taking the lock on the assumption that this will allow better
code movement.

Change-Id: Idd8ba65c0d45c7ceea42d50ea3a6a521e0123733
Signed-off-by: Robert Varga <rovarga@cisco.com>