Update patch set 1
Patch Set 1:

Please ignore my 2) concern in above comment, because as we already send group/meter feature request to the switch whenever it connects to the controller. But those requests were not entertained by the statistics manager, because of following condition in onMeterFeatureNotification (similar in grou feature notification)
 if ( ! isExpectedStatistics(transId, nodeId)) {
            LOG.debug("STAT-MANAGER - MeterFeaturesUpdated: unregistred notification detect TransactionId {}", transId);
            return;
        }
Now anyways we are not planning to send these multipart request from statistics manager, then we should remove this condition, so that statistics-manager can entertain the group/meter features response for the request sent by the plugin at the time of connection.

Patch-set: 1