BUG 2463 - Changing setting for shard-snapshot-data-threshold-percentage does not work
This was missed by the commit which introduced this setting. It was not read from the config
and set on the DataStoreContext. This meant that the default (12 percent) was used regardless
of the actual setting.

Also changed the name of the stat to InMemoryJournalDataSize instead of just DataSize to be
more clear

Change-Id: Iaffddb3e614db7a270d847ba219801e36c8a64e0
Signed-off-by: Moiz Raja <moraja@cisco.com>