Update patch set 2
Patch Set 2:

This patch contains optimizations for the single node cluster case. You can try them out by setting the experimental property to true. The focus of the patches is to minimize usage of an in-memory journal and if both replication and persistence are not enabled to not even serialize the modification.

Do not merge this patch. If and when we decide these look good I will make them available without the experimental flag on and with more testing to go with it.

Patch-set: 2
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0