BUG-2510: Remove all registrations when a routed rpc is closed
The API contract specifies that the effects of a Registration are
removed on close(). Unfortunately the routed RPC case is more funny, as
it has sub-registrations, for each added path -- and we have no mention
of the fact that the user should be removing them.

Hence we need to handle the user expecting us to remove any and all
paths which have been registered.

Change-Id: Iedf451c2a481d648173819f895cccb9e63c54d99
Signed-off-by: Robert Varga <rovarga@cisco.com>