BUG 2509 : Removing all journal entries from a Followers in-memory journal causes Leader to send an InstallSnapshot
This patch simply checks if an AppendEntries#prevLogIndex is equal to the followers ReplicatedLog#snapshotIndex. If
both are the same it also verifies that the AppendEntries#prevLogTerm is equal to the followers ReplicatedLog#snapshotTerm.
This is sufficient in most cases to prevent a re-installation of a snapshot when the AppendEntries message is supposed
to immediately follow the snapshot.

Change-Id: Iab9b493455538544efd06334329f91cc426d61f8
Signed-off-by: Moiz Raja <moraja@cisco.com>