Bug 2526: Race condition may cause missing routes in RPC BucketStore
Created AddToLocalRouter and RemoveFromLocalRouter messages and
changed the BucketStore to update its localRouter RoutingTable on receipt of
the new messages. These replace the GetLocalRouter and UpdateRouter messages.

When updating, the BucketStore makes a local copy of the
RoutingTable which is modified and set in the BucketImpl instance.
This is done to ensure clients always get an immutable RoutingTable
instance when the local BucketImpl is returned in messages. In addition, an
immutable instance of the local BucketImpl is always returned.

The RoutingTable class was converted to an immutable interface with
a mutable implementation, RoutingTableImpl.

Change-Id: Ief96c28775b729f459d324971403222e5a578029
Signed-off-by: tpantelis <tpanteli@brocade.com>