Bug 2526: Race condition may cause missing routes in RPC BucketStore
Changed the RpcRegistry class to derive from BucketStore instead of
creating the BucketStore as a separate actor. This was done because the
copy-on-write update operations in the BucketStore need to be done
atomically so the RpcRegistry needs direct access to the BucketStore in
order to do this safely and efficiently. The RpcRegistry handles the
sematices of the Bucket data - this keeps the BucketStore data agnostic.

When updating, the RpcRegistry makes a local copy of the RoutingTable
which is modified and set in the BucketImpl instance. This is done to
ensure clients always get an immutable RoutingTable instance when the
local BucketImpl is returned in messages. In addition, an immutable instance
of the local BucketImpl is always returned.

The RoutingTable class was converted to an immutable interface with
a mutable implementation, RoutingTableImpl.

Change-Id: Ief96c28775b729f459d324971403222e5a578029
Signed-off-by: tpantelis <tpanteli@brocade.com>