Bug 1234:
For nodes and node-connectors not following the OF format (which is long
for node and short for node-connectors) we were getting lot of
(number-format) exceptions at the SAL-Compatibility layer. (Note that
there wasn't any functionality impact on the MD-SAL applications).

Now based on whether the passed node and node-connectors are OF or
MD_SAL type, these are handled properly.

Signed-off-by: Rahul Sharma <rahushar@ciena.com>
Change-Id: Ia9de0719c0c71eeaa26ce719fd55b483963209ee