Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

There were a couple places in the code where a variable was assigned and then used only once. Not sure if this was for code clarity or another reasons. I personally tend to try to avoid that, but I also saw this pattern in the original code and I suspect the optimizer will take care of it. Marking these as final might help the optimizer a bit more.

Patch-set: 1
Reviewer: Gerrit User 1045 <1045@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1