Optimize transaction submission when backend is idle
If the backend is idle we need to submit the transaction without taking
a round-trip to readyTx, saving us two compare-and-swaps by realizing
this is the case before we publish the transaction.

This is a trade-off, as we pay with an additional volatile read in the
case when the backend is committing a transaction.

Change-Id: Ie0f4f82e607de6462befa76c8b2cd999cb2b0b23
Signed-off-by: Robert Varga <rovarga@cisco.com>