Update patch set 1
Patch Set 1:

What is the intended outcome if the tryAcquire() fails?  Right now it looks like it just logs a warning, and will still proceed with the operation and with the permit release.  Is it a problem if Semaphore.release() is called without having the permit acquired?

Similarly, the ready() method seems to acquire at most one permit but can potentially release multiple permits depending on the number of shards in the cohort.

Patch-set: 1