Update patch set 1
Patch Set 1:

If we release a permit without acquiring it first, it means that the semaphore ends up with extra permits, correct?  It seems that then the ability to throttle would be lost once the number of available permits exceeds actorContext.getTransactionOutstandingOperationLimit().  Are we saying that this doesn't matter because by then the transaction would have failed anyway?

I feel that if we're to use a semaphore, then we should have a one-to-one match between acquire and release as is the usual expectation.  Alternatively, we should have comments in the code explaining why it's safe to release without acquire here.

Patch-set: 1