Update patch set 1
Patch Set 1:

If we cannot acquire permits in the operation timeout window then yes the transaction is about to fail anyway because one of the previous operations is taking too much time to complete. An alternate could be to throw an exception to trigger the failure earlier. It would have to be a runtime exception though so API consumers would not be prepared to handle it. As the API currently stands I would have to wait for the transaction to fail. 

An exception to this rule may be reads but I'm not sure I want to treat this differently. 

Maybe a comment would help or some more debug logging?

Patch-set: 1