Update patch set 1
Patch Set 1:

I wonder if it might be better to create a semaphore per TransactionContext instead of one shared semaphore. This would eliminate the unbalanced acquire/release semantics in "ready" with multiple cohorts. With multiple shards, we might get unnecessary blocking if the total # of operations across all shards exceeds the limit. Also if we can't create a Tx for a shard then we get a NoopTransactionContext but it doesn't release the semaphore which may result in unwanted blocking when trying to acquire. We can make NoopTransactionContext  release but I don't think we need to throttle in this case. What we're really trying to do is throttle operations to the Tx actor(s).

We could call acquire instead of tryAcquire so the caller waits indefinitely. This would also eliminate potential unbalanced acquire/release semantics. With a 5 sec time out on tryAcquire we still could get dropped messages sent to dead letters - waiting indefinitely would/should avoid this possibility.

Patch-set: 1