Update patch set 1
Patch Set 1:

Actually, would it be possible to manage the back pressure using Akka's built-in mechanisms instead of rolling our own?  I would think that we're using a BoundedMailbox precisely so that we don't have to manually throttle the incoming messages ourselves.

For example, what if we just increased Akka's push timeout to 5 seconds?

Patch-set: 1