Update patch set 1
Patch Set 1:

Gary, that won't work (in fact it does not and I have tried it). This is because in-effect we maintain an operations queue on the transaction proxy side and only send the operations to the remote transaction actor when the transaction is ready without blocking the client thread.

Here is how things happen,
Because we really do not know which shard the transaction needs to be created on we wait for the first operation to happen on the transaction. At that point we do know which shard we need to create the transaction on so we go out and find the primary replica for that shard and then we create the transaction with that remote shard. In the meantime the consumer which was never blocked manages to add more operations on the transaction than the bounded mailbox permits. Remember that because we queue the operations and use the akka dispatcher on the transaction proxy side to actually execute those operations we never get to block the client and therefore the akka backpressure does not actually slow down the client.

One solution may be to simply block the client till the previous operation completes. While this seems counter-intuitive it may actually work better than the non-blocking code we have on the transaction proxy side today.

Patch-set: 1