Update patch set 1
Patch Set 1:

Yeah - it would be easier to just block on very operation but I don't think that would be ideal. With the back pressure throttling we're only blocking when there's a flood of operations in  short period of time, which should be relatively uncommon.

I think implementing throttling via our own bounded queue would add significant complexity to the code. We would need a thread to pull from the queue so we'd essentially be re-implmenting what akka already does. I think the use of a Semaphore is a simple, elegant way to gain throttling.

Patch-set: 1