Update patch set 1
Patch Set 1:

Tom, this is a revisit of an older conversation we had - what if we introduced a transaction actor on the client side which would queue up all the messages for a given transaction. When handling any message in there we could block till the reply came back. I think this would help to eliminate our TransactionProxy queuing and then we would be able to use the akka push timeout as well. What do you think? Do you see any problems with that?

Patch-set: 1