Update patch set 1
Patch Set 1:

The reason I suggested the bounded queue is that we're effectively implementing a bounded queue using a semaphore (and having to deal with acquire/release matching issues, etc.), so why not use a pre-canned one that already takes care of that?  We can use a bounded queue without needing a new thread to do the polling, in the same way that we didn't need a new thread for the semaphore implementation.

That said, I think Moiz's suggestion to make the TransactionProxy front-end actor-based is the better idea.  That thought had occurred to me earlier when I was refactoring this class; I assume there were  reasons that this was not done originally?

Patch-set: 1