Update patch set 1
Patch Set 1:

We discussed the separate actor on the hackers call today. My concern is that adding another actor in the chain might degrade performance. It's well know that a major bottleneck in async programming are blocking queues and context switching. Adding another actor would add another blocking queue transfer. We decided to go ahead with this patch and revisit this later.  

Re: the bounded queue, we only batch operations until the Tx actor is created - after that we send directly to the actor. We don't explicitly create a thread to execute the batch - we get a thread for "free" from akka via the Future callback.

Patch-set: 1