Update patch set 1
Patch Set 1:

So to finish this patch, we should also do the permit release in the NoOpTransactionContext.

Re: the ready() method, we could acquire permits for each shard sub-Tx but I don't think it really matters either way. ready() is called only once and is the last operation a client can (legally) perform on the Tx.

Patch-set: 1