Change trackerList to a LinkedList
The trackers are being added to the end and usually remove in a linear
fashion. An ArrayList's remove() method makes sure the backing array is
packed -- which means we end up copying the array over and over again.

Changing the trackerList to a LinkedList removes the need for copying at
the expense of memory footprint and cache locality.

Change-Id: I3e40487798eb64713ed4526f53a0a8c385a9a778
Signed-off-by: Robert Varga <rovarga@cisco.com>